+++ CHANGELOG.md
@@ -13,11 +13,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added "IP Cores", "Externals", and "Constants" layers to the GUI which you can hide/show from the settings
 - Nox session for downloading and packaging FuseSoc libraries
 - Support Python 3.13 and dropped support for Python 3.8
+- Added to config option for choosing a location where KPM server should be built
 - All in one command `topwrap gui` for building, starting the KPM server and connecting the client to it.
 
 ### Changed
 
 - All YAML files internal to Topwrap (examples, tests, built-in repo) now only use the `.yaml` extension
+- Default location for built KPM server from `./build/` to `$XDG_CACHE_HOME/topwrap/kpm_build`
 
 ### Fixed